refactor wip: restore global entity selector
