+++ baker/formatWordpressPost.tsx
@@ -18,7 +18,7 @@ import {
     JsonError,
     TocHeading,
 } from "../clientUtils/owidTypes"
-import { bakeGlobalEntitySelector } from "./bakeGlobalEntitySelector"
+import { bakeGlobalEntitySelector } from "../site/bakeGlobalEntitySelector"
 import { Footnote } from "../site/Footnote"
 import { LoadingIndicator } from "../grapher/loadingIndicator/LoadingIndicator"
 import { PROMINENT_LINK_CLASSNAME } from "../site/blocks/ProminentLink"
@@ -563,10 +563,6 @@ export const formatWordpressPost = async (
             )
         }
     })
-
-    // Render global country selection component.
-    // Injects a <section>, which is why it executes last.
-    bakeGlobalEntitySelector(cheerioEl)
 
     return {
         id: post.id,
