+++ site/bakeGlobalEntitySelector.tsx
@@ -8,7 +8,9 @@ import {
 } from "../grapher/controls/globalEntitySelector/GlobalEntitySelectorConstants"
 import { SelectionArray } from "../grapher/selection/SelectionArray"
 
-export const bakeGlobalEntitySelector = (cheerioEl: CheerioStatic) => {
+export const bakeGlobalEntitySelector = (
+    cheerioEl: CheerioStatic
+): CheerioStatic => {
     // The data attr used to be `data-entity-select`, but later changed for consistency in the code.
     // But we should still support the old attribute.
     cheerioEl(`*[data-entity-select], ${GLOBAL_ENTITY_SELECTOR_ELEMENT}`).each(
@@ -31,4 +33,6 @@ export const bakeGlobalEntitySelector = (cheerioEl: CheerioStatic) => {
             $el.html(rendered).insertAfter($section)
         }
     )
+
+    return cheerioEl
 }
